# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PiXELL-River Financial",
    "version": "1.0.0",
    "description": "Managing employees and branches"
  },
  "paths": {
    "/api/v1/branches": {
      "get": {
        "summary": "Retrieve a list of all branches",
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "A list of branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new branch",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new branch.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Branch created successfully."
                    },
                    "data": {
                      "$ref": "#/components/schemas/Branch"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/branches/{id}": {
      "get": {
        "summary": "Retrieve a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "description": "The unique identifier of the branch."
          }
        ],
        "responses": {
          "200": {
            "description": "Branch details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "put": {
        "summary": "Update a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "description": "The unique identifier of the branch."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated branch details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "description": "The unique identifier of the branch."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the branch.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Branch deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    },
    "/api/v1/employees": {
      "get": {
        "summary": "Retrieve a list of all employees",
        "tags": [
          "Employees"
        ],
        "responses": {
          "200": {
            "description": "A list of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new employee.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Employee created successfully."
                    },
                    "data": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "summary": "Retrieve an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "12345"
            },
            "description": "The unique identifier of the employee."
          }
        ],
        "responses": {
          "200": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "put": {
        "summary": "Update an employee's details by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "12345"
            },
            "description": "The unique identifier of the employee."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated employee details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "12345"
            },
            "description": "The unique identifier of the employee."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the employee.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Employee deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    },
    "/api/v1/employees/branch/{branchId}": {
      "get": {
        "summary": "Retrieve all employees for a specific branch",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "description": "The ID of the branch."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees for the branch.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No employees found for the branch."
          }
        }
      }
    },
    "/api/v1/employees/department/{department}": {
      "get": {
        "summary": "Retrieve all employees for a specific department",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "department",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Management"
            },
            "description": "The name of the department."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees for the department.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No employees found for the department."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Branch": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string",
            "description": "The unique identifier for a branch."
          },
          "branchName": {
            "type": "string",
            "description": "The name of the branch."
          },
          "branchAddress": {
            "type": "string",
            "description": "The physical address of the branch."
          },
          "branchPhone": {
            "type": "string",
            "description": "The contact phone number for the branch."
          }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "description": "The unique identifier for an employee."
          },
          "employeeName": {
            "type": "string",
            "description": "The full name of the employee."
          },
          "employeePosition": {
            "type": "string",
            "description": "The employee's job title."
          },
          "employeeDepartment": {
            "type": "string",
            "description": "The department where the employee works."
          },
          "employeeEmail": {
            "type": "string",
            "format": "email",
            "description": "The employee's email address."
          },
          "employeePhone": {
            "type": "string",
            "description": "The employee's phone number."
          },
          "employeeBranchId": {
            "type": "string",
            "description": "The ID of the branch the employee is associated with."
          }
        }
      }
    }
  },
  "tags": []
}